requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
RPi.GPIO==0.7.1
simple-websocket==1.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
